requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.8
load-dotenv==0.1.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.2
